add serialization tests part 1
