+++ test/integration/backward_compatible/sql/RowTest.js
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2008-2021, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+'use strict';
+
+const RC = require('../../RC');
+const TestUtil = require('../../../TestUtil');
+const { Client } = require('../../../../');
+
+describe('SqlRowTest', function () {
+    let client;
+    let cluster;
+    let someMap;
+    let mapName;
+
+    before(async function () {
+        TestUtil.markClientVersionAtLeast(this, '4.2');
+        cluster = await RC.createCluster(null, null);
+        await RC.startMember(cluster.id);
+        client = await Client.newHazelcastClient({
+            clusterName: cluster.id
+        });
+    });
+
+    beforeEach(async function () {
+        mapName = TestUtil.randomString(10);
+        someMap = await client.getMap(mapName);
+    });
+
+    after(async function () {
+        await RC.terminateCluster(cluster.id);
+        await client.shutdown();
+    });
+
+    afterEach(async function () {
+        await someMap.clear();
+    });
+
+});
